Move helper code into Util class
